+++ tests/src/interfaces/default/types.ts
@@ -210,21 +210,44 @@ export interface CumulusPalletXcmpQueueError extends Enum {
 /** @name CumulusPalletXcmpQueueEvent */
 export interface CumulusPalletXcmpQueueEvent extends Enum {
   readonly isSuccess: boolean;
-  readonly asSuccess: Option<H256>;
+  readonly asSuccess: {
+    readonly messageHash: Option<H256>;
+    readonly weight: u64;
+  } & Struct;
   readonly isFail: boolean;
-  readonly asFail: ITuple<[Option<H256>, XcmV2TraitsError]>;
+  readonly asFail: {
+    readonly messageHash: Option<H256>;
+    readonly error: XcmV2TraitsError;
+    readonly weight: u64;
+  } & Struct;
   readonly isBadVersion: boolean;
-  readonly asBadVersion: Option<H256>;
+  readonly asBadVersion: {
+    readonly messageHash: Option<H256>;
+  } & Struct;
   readonly isBadFormat: boolean;
-  readonly asBadFormat: Option<H256>;
+  readonly asBadFormat: {
+    readonly messageHash: Option<H256>;
+  } & Struct;
   readonly isUpwardMessageSent: boolean;
-  readonly asUpwardMessageSent: Option<H256>;
+  readonly asUpwardMessageSent: {
+    readonly messageHash: Option<H256>;
+  } & Struct;
   readonly isXcmpMessageSent: boolean;
-  readonly asXcmpMessageSent: Option<H256>;
+  readonly asXcmpMessageSent: {
+    readonly messageHash: Option<H256>;
+  } & Struct;
   readonly isOverweightEnqueued: boolean;
-  readonly asOverweightEnqueued: ITuple<[u32, u32, u64, u64]>;
+  readonly asOverweightEnqueued: {
+    readonly sender: u32;
+    readonly sentAt: u32;
+    readonly index: u64;
+    readonly required: u64;
+  } & Struct;
   readonly isOverweightServiced: boolean;
-  readonly asOverweightServiced: ITuple<[u64, u64]>;
+  readonly asOverweightServiced: {
+    readonly index: u64;
+    readonly used: u64;
+  } & Struct;
   readonly type: 'Success' | 'Fail' | 'BadVersion' | 'BadFormat' | 'UpwardMessageSent' | 'XcmpMessageSent' | 'OverweightEnqueued' | 'OverweightServiced';
 }
 
@@ -708,17 +731,17 @@ export interface FrameSystemPhase extends Enum {
 
 /** @name OpalRuntimeOriginCaller */
 export interface OpalRuntimeOriginCaller extends Enum {
-  readonly isVoid: boolean;
-  readonly asVoid: SpCoreVoid;
   readonly isSystem: boolean;
   readonly asSystem: FrameSupportDispatchRawOrigin;
+  readonly isVoid: boolean;
+  readonly asVoid: SpCoreVoid;
   readonly isPolkadotXcm: boolean;
   readonly asPolkadotXcm: PalletXcmOrigin;
   readonly isCumulusXcm: boolean;
   readonly asCumulusXcm: CumulusPalletXcmOrigin;
   readonly isEthereum: boolean;
   readonly asEthereum: PalletEthereumRawOrigin;
-  readonly type: 'Void' | 'System' | 'PolkadotXcm' | 'CumulusXcm' | 'Ethereum';
+  readonly type: 'System' | 'Void' | 'PolkadotXcm' | 'CumulusXcm' | 'Ethereum';
 }
 
 /** @name OpalRuntimeRuntime */
@@ -1567,6 +1590,17 @@ export interface PalletTimestampCall extends Enum {
   readonly type: 'Set';
 }
 
+/** @name PalletTransactionPaymentEvent */
+export interface PalletTransactionPaymentEvent extends Enum {
+  readonly isTransactionFeePaid: boolean;
+  readonly asTransactionFeePaid: {
+    readonly who: AccountId32;
+    readonly actualFee: u128;
+    readonly tip: u128;
+  } & Struct;
+  readonly type: 'TransactionFeePaid';
+}
+
 /** @name PalletTransactionPaymentReleases */
 export interface PalletTransactionPaymentReleases extends Enum {
   readonly isV1Ancient: boolean;
@@ -1589,11 +1623,16 @@ export interface PalletTreasuryCall extends Enum {
   readonly asApproveProposal: {
     readonly proposalId: Compact<u32>;
   } & Struct;
+  readonly isSpend: boolean;
+  readonly asSpend: {
+    readonly amount: Compact<u128>;
+    readonly beneficiary: MultiAddress;
+  } & Struct;
   readonly isRemoveApproval: boolean;
   readonly asRemoveApproval: {
     readonly proposalId: Compact<u32>;
   } & Struct;
-  readonly type: 'ProposeSpend' | 'RejectProposal' | 'ApproveProposal' | 'RemoveApproval';
+  readonly type: 'ProposeSpend' | 'RejectProposal' | 'ApproveProposal' | 'Spend' | 'RemoveApproval';
 }
 
 /** @name PalletTreasuryError */
@@ -1601,8 +1640,9 @@ export interface PalletTreasuryError extends Enum {
   readonly isInsufficientProposersBalance: boolean;
   readonly isInvalidIndex: boolean;
   readonly isTooManyApprovals: boolean;
+  readonly isInsufficientPermission: boolean;
   readonly isProposalNotApproved: boolean;
-  readonly type: 'InsufficientProposersBalance' | 'InvalidIndex' | 'TooManyApprovals' | 'ProposalNotApproved';
+  readonly type: 'InsufficientProposersBalance' | 'InvalidIndex' | 'TooManyApprovals' | 'InsufficientPermission' | 'ProposalNotApproved';
 }
 
 /** @name PalletTreasuryEvent */
@@ -1638,7 +1678,13 @@ export interface PalletTreasuryEvent extends Enum {
   readonly asDeposit: {
     readonly value: u128;
   } & Struct;
-  readonly type: 'Proposed' | 'Spending' | 'Awarded' | 'Rejected' | 'Burnt' | 'Rollover' | 'Deposit';
+  readonly isSpendApproved: boolean;
+  readonly asSpendApproved: {
+    readonly proposalIndex: u32;
+    readonly amount: u128;
+    readonly beneficiary: AccountId32;
+  } & Struct;
+  readonly type: 'Proposed' | 'Spending' | 'Awarded' | 'Rejected' | 'Burnt' | 'Rollover' | 'Deposit' | 'SpendApproved';
 }
 
 /** @name PalletTreasuryProposal */
@@ -1684,7 +1730,7 @@ export interface PalletUniqueCall extends Enum {
   readonly isAddCollectionAdmin: boolean;
   readonly asAddCollectionAdmin: {
     readonly collectionId: u32;
-    readonly newAdmin: PalletEvmAccountBasicCrossAccountIdRepr;
+    readonly newAdminId: PalletEvmAccountBasicCrossAccountIdRepr;
   } & Struct;
   readonly isRemoveCollectionAdmin: boolean;
   readonly asRemoveCollectionAdmin: {