+++ runtime/opal/CHANGELOG.md
@@ -1,3 +1,18 @@
 # Change Log
 
 All notable changes to this project will be documented in this file.
+
+<!-- bureaucrate goes here -->
+## [v0.9.27] 2022-08-16
+
+### Bugfixes
+
+- Add missing config keys 74f532ac28dce15c15e7d576c074a58eba658c08
+
+### Other changes
+
+- build: Upgrade polkadot to v0.9.27 2c498572636f2b34d53b1c51b7283a761a7dc90a
+
+- build: Upgrade polkadot to v0.9.26 85515e54c4ca1b82a2630034e55dcc804c643bf8
+
+- build: Upgrade polkadot to v0.9.25 cdfb9bdc7b205ff1b5134f034ef9973d769e5e6b